# Conflicts:
#	server/src/main/java/org/apache/iotdb/db/metadata/MManager.java

+++ server/src/main/java/org/apache/iotdb/db/engine/storagegroup/TsFileProcessor.java
@@ -271,9 +271,9 @@ public class TsFileProcessor {
    * the range [start, end)
    *
    * @param insertTabletPlan insert a tablet of a device
-   * @param start start index of rows to be inserted in insertTabletPlan
-   * @param end end index of rows to be inserted in insertTabletPlan
-   * @param results result array
+   * @param start            start index of rows to be inserted in insertTabletPlan
+   * @param end              end index of rows to be inserted in insertTabletPlan
+   * @param results          result array
    */
   public void insertTablet(InsertTabletPlan insertTabletPlan, int start, int end,
       TSStatus[] results) throws WriteProcessException {
@@ -809,7 +809,11 @@ public class TsFileProcessor {
         }
         if (targetFile.getName().endsWith(TSFILE_SUFFIX)) {
           if (!isMergeFinished) {
-            writer.getIOWriterOut().truncate(offset - 1);
+            logger.info("{}: {} merge recover {} level vms to TsFile", storageGroupName,
+                tsFileResource.getTsFile().getName(), vmWriters.size());
+            if (offset > 0) {
+              writer.getIOWriterOut().truncate(offset - 1);
+            }
             VmMergeUtils.merge(writer, packVmWritersToSequenceList(vmWriters),
                 storageGroupName,
                 new VmLogger(tsFileResource.getTsFile().getParent(),
@@ -833,32 +837,33 @@ public class TsFileProcessor {
               newVmWriter.setFile(newVmFile);
             }
           } else {
-            if (deviceSet.isEmpty()) {
-              Files.delete(targetFile.toPath());
-            } else {
+            logger.info("{}: {} [Hot Compaction Recover] merge level-{}'s {} vms to next level vm",
+                storageGroupName, tsFileResource.getTsFile().getName(), level,
+                sourceFileList.size());
+            if (offset > 0) {
               newVmWriter.getIOWriterOut().truncate(offset - 1);
-              // vm files must be sequence, so we just have to find the first file
-              int startIndex = 0;
-              for (startIndex = 0; startIndex < vmWriters.get(level).size(); startIndex++) {
-                RestorableTsFileIOWriter levelVmWriter = vmWriters.get(level).get(startIndex);
-                if (levelVmWriter.getFile().getAbsolutePath()
-                    .equals(sourceFileList.get(0).getAbsolutePath())) {
-                  break;
-                }
+            }
+            // vm files must be sequence, so we just have to find the first file
+            int startIndex;
+            for (startIndex = 0; startIndex < vmWriters.get(level).size(); startIndex++) {
+              RestorableTsFileIOWriter levelVmWriter = vmWriters.get(level).get(startIndex);
+              if (levelVmWriter.getFile().getAbsolutePath()
+                  .equals(sourceFileList.get(0).getAbsolutePath())) {
+                break;
               }
-              List<RestorableTsFileIOWriter> levelVmWriters = new ArrayList<>(
-                  vmWriters.get(level).subList(startIndex, startIndex + sourceFileList.size()));
-              List<TsFileResource> levelVmFiles = new ArrayList<>(
-                  vmTsFileResources.get(level)
-                      .subList(startIndex, startIndex + sourceFileList.size()));
-              VmMergeUtils.merge(newVmWriter, levelVmWriters,
-                  storageGroupName,
-                  new VmLogger(tsFileResource.getTsFile().getParent(),
-                      tsFileResource.getTsFile().getName()),
-                  deviceSet, sequence);
-              for (int i = 0; i < vmWriters.size(); i++) {
-                deleteVmFiles(levelVmFiles, levelVmWriters);
-              }
+            }
+            List<RestorableTsFileIOWriter> levelVmWriters = new ArrayList<>(
+                vmWriters.get(level).subList(startIndex, startIndex + sourceFileList.size()));
+            List<TsFileResource> levelVmFiles = new ArrayList<>(
+                vmTsFileResources.get(level)
+                    .subList(startIndex, startIndex + sourceFileList.size()));
+            VmMergeUtils.merge(newVmWriter, levelVmWriters,
+                storageGroupName,
+                new VmLogger(tsFileResource.getTsFile().getParent(),
+                    tsFileResource.getTsFile().getName()),
+                deviceSet, sequence);
+            for (int i = 0; i < vmWriters.size(); i++) {
+              deleteVmFiles(levelVmFiles, levelVmWriters);
             }
           }
         }
@@ -895,8 +900,8 @@ public class TsFileProcessor {
               tsFileResource.getTsFile().getName());
           File newVmFile = createNewVMFileWithLock(tsFileResource, 0);
           if (vmWriters.isEmpty()) {
-            vmWriters.add(new ArrayList<>());
-            vmTsFileResources.add(new ArrayList<>());
+            vmWriters.add(new CopyOnWriteArrayList<>());
+            vmTsFileResources.add(new CopyOnWriteArrayList<>());
           }
           vmTsFileResources.get(0).add(new TsFileResource(newVmFile));
           curWriter = new RestorableTsFileIOWriter(newVmFile);
@@ -907,7 +912,7 @@ public class TsFileProcessor {
           curWriter = writer;
         }
         curWriter.mark();
-        flushTask = new MemTableFlushTask(memTableToFlush, curWriter, storageGroupName);
+        flushTask = new MemTableFlushTask(memTableToFlush, curWriter, storageGroupName, writer);
         flushTask.syncFlushMemTable();
       } catch (Exception e) {
         logger.error("{}: {} meet error when flushing a memtable, change system mode to read-only",
@@ -945,37 +950,41 @@ public class TsFileProcessor {
 
     if (shouldClose && flushingMemTables.isEmpty()) {
       try {
-        // merge vm to tsfile
-        while (true) {
-          if (!mergeWorking) {
-            break;
+        if (config.isEnableVm()) {
+          // merge vm to tsfile
+          while (true) {
+            if (!mergeWorking) {
+              break;
+            }
+            try {
+              TimeUnit.MILLISECONDS.sleep(10);
+            } catch (@SuppressWarnings("squid:S2142") InterruptedException e) {
+              logger.error("{}: {}, closing task is interrupted.",
+                  storageGroupName, tsFileResource.getTsFile().getName(), e);
+              // generally it is because the thread pool is shutdown so the task should be aborted
+              break;
+            }
           }
-          try {
-            TimeUnit.MILLISECONDS.sleep(10);
-          } catch (@SuppressWarnings("squid:S2142") InterruptedException e) {
-            logger.error("{}: {}, closing task is interrupted.",
-                storageGroupName, tsFileResource.getTsFile().getName(), e);
-            // generally it is because the thread pool is shutdown so the task should be aborted
-            break;
+          logger.info("{}: [Hot Compaction] Start to merge total {} levels' vm to TsFile {}",
+              storageGroupName, vmTsFileResources.size() + 1, tsFileResource.getTsFile().getName());
+          long startTimeMillis = System.currentTimeMillis();
+          VmLogger vmLogger = new VmLogger(tsFileResource.getTsFile().getParent(),
+              tsFileResource.getTsFile().getName());
+          vmLogger.logFile(TARGET_NAME, writer.getFile());
+          flushAllVmToTsFile(vmWriters, vmTsFileResources, vmLogger);
+          vmLogger.logMergeFinish();
+          vmLogger.close();
+          File logFile = FSFactoryProducer.getFSFactory()
+              .getFile(tsFileResource.getTsFile().getParent(),
+                  tsFileResource.getTsFile().getName() + VM_LOG_NAME);
+          if (logFile.exists()) {
+            Files.delete(logFile.toPath());
           }
+          logger
+              .info("{}: [Hot Compaction] All vms are merged to TsFile {}, time consumption: {} ms",
+                  storageGroupName, tsFileResource.getTsFile().getName(),
+                  System.currentTimeMillis() - startTimeMillis);
         }
-        logger.info("{}: [Hot Compaction] Start to merge total {} levels' vm to TsFile {}",
-            storageGroupName, vmTsFileResources.size() + 1, tsFileResource.getTsFile().getName());
-        long startTimeMillis = System.currentTimeMillis();
-        VmLogger vmLogger = new VmLogger(tsFileResource.getTsFile().getParent(),
-            tsFileResource.getTsFile().getName());
-        flushAllVmToTsFile(vmWriters, vmTsFileResources, vmLogger);
-        vmLogger.logMergeFinish();
-        vmLogger.close();
-        File logFile = FSFactoryProducer.getFSFactory()
-            .getFile(tsFileResource.getTsFile().getParent(),
-                tsFileResource.getTsFile().getName() + VM_LOG_NAME);
-        if (logFile.exists()) {
-          Files.delete(logFile.toPath());
-        }
-        logger.info("{}: [Hot Compaction] All vms are merged to TsFile {}, time consumption: {} ms",
-            storageGroupName, tsFileResource.getTsFile().getName(),
-            System.currentTimeMillis() - startTimeMillis);
         writer.mark();
         try {
           double compressionRatio = ((double) totalMemTableSize) / writer.getPos();
@@ -1029,7 +1038,7 @@ public class TsFileProcessor {
       }
     } else {
       // on other merge task working now, it's safe to submit one.
-      if (!mergeWorking) {
+      if (config.isEnableVm() && !mergeWorking) {
         mergeWorking = true;
         logger.info("{}: {} submit a vm merge task", storageGroupName,
             tsFileResource.getTsFile().getName());
@@ -1119,10 +1128,10 @@ public class TsFileProcessor {
    * memtables and then compact them into one TimeValuePairSorter). Then get the related
    * ChunkMetadata of data on disk.
    *
-   * @param deviceId device id
+   * @param deviceId      device id
    * @param measurementId measurements id
-   * @param dataType data type
-   * @param encoding encoding
+   * @param dataType      data type
+   * @param encoding      encoding
    */
   public void query(String deviceId, String measurementId, TSDataType dataType, TSEncoding encoding,
       Map<String, String> props, QueryContext context,
@@ -1176,18 +1185,9 @@ public class TsFileProcessor {
         for (int j = 0; j < vmWriters.get(i).size(); j++) {
           RestorableTsFileIOWriter vmWriter = vmWriters.get(i).get(j);
           TsFileResource vmTsFileResource = vmTsFileResources.get(i).get(j);
-          for (Entry<String, Map<String, List<ChunkMetadata>>> entry : vmWriter
-              .getMetadatasForQuery()
-              .entrySet()) {
-            String device = entry.getKey();
-            for (List<ChunkMetadata> tmpChunkMetadataList : entry.getValue().values()) {
-              for (ChunkMetadata chunkMetadata : tmpChunkMetadataList) {
-                vmTsFileResource.updateStartTime(device, chunkMetadata.getStartTime());
-                if (!sequence) {
-                  vmTsFileResource.updateEndTime(device, chunkMetadata.getEndTime());
-                }
-              }
-            }
+          vmTsFileResource.updateStartTime(deviceId, tsFileResource.getStartTime(deviceId));
+          if (!sequence) {
+            vmTsFileResource.updateEndTime(deviceId, tsFileResource.getEndTime(deviceId));
           }
           chunkMetadataList = vmWriter.getVisibleMetadataList(deviceId, measurementId, dataType);
           QueryUtils.modifyChunkMetaData(chunkMetadataList,
@@ -1281,8 +1281,8 @@ public class TsFileProcessor {
         long vmPointNum = 0;
         // all flush to target file
         Map<Path, MeasurementSchema> pathMeasurementSchemaMap = new HashMap<>();
-        for (List<RestorableTsFileIOWriter> subVmWriters : vmMergeWriters) {
-          for (RestorableTsFileIOWriter vmWriter : subVmWriters) {
+        for (List<RestorableTsFileIOWriter> levelVmWriters : vmMergeWriters) {
+          for (RestorableTsFileIOWriter vmWriter : levelVmWriters) {
             Map<String, Map<String, List<ChunkMetadata>>> schemaMap = vmWriter
                 .getMetadatasForQuery();
             for (Entry<String, Map<String, List<ChunkMetadata>>> schemaMapEntry : schemaMap
